[app-5]: gradle + console spring-boot
