+++ .gitignore
@@ -6,6 +6,10 @@ out/
 .idea
 *.iml
 
+gradlew
+gradlew.bat
+.gradle
+gradle/*
 
 # Log file
 *.log
@@ -27,3 +31,8 @@ out/
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
